+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * f790c1f - 2017-10-09: Fixed issues with API multi commands 
+- * fc737f8 - 2017-10-09: Added option delete to post-process cmd for API 
+- * bd5e378 - 2017-10-09: Fixed issues with API multi commands 
 - * e4be7cf - 2017-10-09: Fixed issue with API help command 
 - * 68b8996 - 2017-10-08: Fixed API builder 
 - * a455d99 - 2017-10-08: Fixed API cmds